requirement.txt
opencv-python==4.4.0
Pillow==8.0.0
numpy==1.19.2
